feat: read expenses using Deno.KV
