+++ types/expenses-tracker/mockExpenses.ts
@@ -1,0 +1,28 @@
+//? Mock expenses
+export const mockExpenses = [
+  {
+    date: new Date(2020, 1, 1),
+    description: "Medical Insurance",
+    cost: 3600,
+  },
+  {
+    date: new Date(2020, 5, 3),
+    description: "Shoes",
+    cost: 200,
+  },
+  {
+    date: new Date(2020, 8, 18),
+    description: "House",
+    cost: 250000,
+  },
+  {
+    date: new Date(2020, 9, 20),
+    description: "Car",
+    cost: 40000,
+  },
+  {
+    date: new Date(2020, 11, 28),
+    description: "Kids",
+    cost: 10000,
+  },
+];
